massive structure revamp, still needs refactoring and restructuring
