+++ build.sh
@@ -1,5 +1,9 @@
 #!/usr/bin/env sh
 
+###################################################
+#Should be installed through git clone --recursive#
+###################################################
+
 ######################################
 #initialize triblers crap            #
 #only works on ubuntu 14.04 and 16.04#
@@ -10,6 +14,10 @@ sudo add-apt-repository "deb http://archive.ubuntu.com/ubuntu vivid main"
 
 # update the apt-get folder.
 sudo apt-get update
+
+# ensure python and pip are installed
+apt-get install -y python
+apt-get install -y python-pip
 
 #install libsodium13 which does not exist anymore under 14.04
 rm libsodium13_1.0.1-1_amd64.deb
@@ -26,9 +34,9 @@ sudo apt-get install -y libav-tools libjs-excanvas libjs-mootools libx11-6 pytho
 pip install decorator==4.0.9
 
 #install tribler itself
-rm tribler_6.5.1_all.deb
-wget https://github.com/Tribler/tribler/releases/download/v6.5.1/tribler_6.5.1_all.deb
-dpkg -i tribler_6.5.1_all.deb
+rm tribler_6.5.2_all.deb
+wget https://github.com/Tribler/tribler/releases/download/v6.5.2/tribler_6.5.2_all.deb
+dpkg -i tribler_6.5.2_all.deb
 apt-get install -y -f
 
 ###########################
@@ -36,19 +44,14 @@ apt-get install -y -f
 ###########################
 apt-get install -y gnome-terminal default-jre
 
-###########################
-#initialize the submodules#
-###########################
-git submodule add https://github.com/Tribler/tribler
-git submodule add https://github.com/weex/addrgen addrgen-master
-#git submodule init --recursive
-#git submodule update --recursive
+##############################
+#Install firefox for selenium#
+##############################
+apt-get install -y firefox
 
-apt-get install -y python
-#apt-get install -y pip3
-
-# Necessities only available through apt-get
-#apt-get install -y firefox
+###################################
+#Install other python requirements#
+###################################
 
 pip install -r requirements.txt
