+++ run.sh
@@ -9,4 +9,4 @@ nohup java -jar selenium-server-standalone-2.53.0.jar &
 #################
 #Start the agent#
 #################
-nohup PYTHONPATH=~/Skynet2.0 python src/agent/agentCore.py &
+python src/agent/agentCore.py
